Add 'break_UNREACHABLE' wrapper
Wrapper for break statements that are unreachable. In some cases,
we prefer to keep them anyway. In those cases, the code above it
is complicated. Changes there may lead to unintentional fall
throughs. This wrapper suppresses the warning on clang.

As an additional check, it adds a 'BUG_ON(1)' when we are in
debug validation mode.

Use this wrapper in the SSL parsing code.