detect-state: handle duplicate inspect/match
If for a packet we have a TX N that has detect state and a TX N+1 that
has no detect state, but does have 'progress', we have a corner case
in stateful detection.

ContinueDetection inspects TX N, but cannot flag the rule in the
de_state_sig_array as the next (TX N+1) has already started and needs
to be inspected. 'StartDetection' however, is then unaware of the fact
that ContinueDetection already inspected the rule. It uses the per
session 'inspect_id' that is only moved forward at the end of the
detection run.

This patch adds a workaround. It creates a new temporary table in
the DetectEngineThreadCtx that ContinueDetection can use to store
the id of TX N. StartDetection then uses this id instead of 'inspect_id'
to start it's detection.

The array is currently quite large, 8 x (number of sigs) bytes.